create .nvim dir with :W

+++ src/lua/util/persist.lua
@@ -8,23 +8,30 @@ function M.getPersistPath()
     return persistPath
 end
 
-local function pathExists(path)
+function M.pathExists(path)
     return io.open(path)~=nil
 end
 
 function M.getWorkspaceFile()
     local dataPath = M.getPersistPath()
-    if not pathExists(dataPath) then
+    if not M.pathExists(dataPath) then
         return nil
     end
     dataPath = dataPath .. '/' .. LAYOUT_FILE
-    if not pathExists(dataPath) then
+    if not M.pathExists(dataPath) then
         return nil
     end
     return dataPath
 end
 
-local function createNvimDir()
+function M.persistPathExists()
+    return M.pathExists(M.getPersistPath())
+end
+
+function M.createNvimDir()
+    if M.persistPathExists() then
+        return
+    end
     local path = M.getPersistPath()
     os.execute('mkdir ' .. path)
     local file = io.open(path .. '/.gitignore', 'w')
@@ -33,11 +40,6 @@ local function createNvimDir()
     end
     file:write(LAYOUT_FILE .. '\n', '\n')
     file:close()
-end
-
--- create .nvim directory on startup if doesn't exist
-if not pathExists(M.getPersistPath()) then
-    createNvimDir()
 end
 
 return M
